refactor(app): remove MuteDetector and mute masking
What: Delete MuteDetector (AX API polling every 0.5s) and the
corresponding applyMuteMask logic in AudioMixer.

Reasoning:
- Problem: MuteDetector polled the accessibility tree at 0.5s intervals
  during every recording, adding steady CPU load even when not in use.
- The feature was only effective for Teams on macOS and required
  Accessibility permission trust — rarely granted in practice.
- Decision: Remove entirely. The echo suppression in AudioMixer already
  handles the most common artifact (hearing yourself in the mix).
