+++ app/MeetingTranscriber/Sources/AXHelper.swift
@@ -1,6 +1,6 @@
 import ApplicationServices
 
-/// Shared Accessibility API helpers used by MuteDetector and ParticipantReader.
+/// Shared Accessibility API helpers used by ParticipantReader.
 enum AXHelper {
     /// Read a single AX attribute value from an element.
     static func getAttribute(_ element: AXUIElement, attribute: String) -> AnyObject? {
