+++ app/MeetingTranscriber/Tests/AudioMixerTests.swift
@@ -38,108 +38,6 @@ final class AudioMixerTests: XCTestCase {
         XCTAssertTrue(result.isEmpty)
     }
 
-    // MARK: - Mute Masking
-
-    func testMuteMaskZeroesSamples() {
-        // 1 second at 100 Hz = 100 samples
-        var samples = [Float](repeating: 1.0, count: 100)
-        let timeline = [
-            MuteTransition(timestamp: 10.0, isMuted: true), // mute at t=10
-            MuteTransition(timestamp: 10.5, isMuted: false), // unmute at t=10.5
-        ]
-
-        AudioMixer.applyMuteMask(
-            samples: &samples,
-            timeline: timeline,
-            sampleRate: 100,
-            micDelay: 0,
-            recordingStart: 10.0,
-        )
-
-        // First 50 samples (0.0s-0.5s) should be zeroed (muted)
-        for i in 0 ..< 50 {
-            XCTAssertEqual(samples[i], 0.0, "Sample \(i) should be muted")
-        }
-        // Remaining should be untouched
-        for i in 50 ..< 100 {
-            XCTAssertEqual(samples[i], 1.0, "Sample \(i) should be untouched")
-        }
-    }
-
-    func testMuteMaskStillMutedAtEnd() {
-        var samples = [Float](repeating: 1.0, count: 100)
-        let timeline = [
-            MuteTransition(timestamp: 5.0, isMuted: true), // mute, never unmuted
-        ]
-
-        AudioMixer.applyMuteMask(
-            samples: &samples,
-            timeline: timeline,
-            sampleRate: 100,
-            recordingStart: 5.0,
-        )
-
-        // All samples should be zeroed
-        XCTAssertTrue(samples.allSatisfy { $0 == 0.0 })
-    }
-
-    func testMuteMaskLargeMicDelayDoesNotCrash() {
-        // micDelay larger than timeline timestamps → range.end < range.start
-        var samples = [Float](repeating: 1.0, count: 100)
-        let timeline = [
-            MuteTransition(timestamp: 10.0, isMuted: true),
-            MuteTransition(timestamp: 10.5, isMuted: false),
-        ]
-
-        AudioMixer.applyMuteMask(
-            samples: &samples,
-            timeline: timeline,
-            sampleRate: 100,
-            micDelay: 20.0,
-            recordingStart: 10.0,
-        )
-
-        // All samples should remain untouched (mute range is entirely negative)
-        XCTAssertTrue(samples.allSatisfy { $0 == 1.0 })
-    }
-
-    func testMuteMaskPartiallyNegativeRange() {
-        // Mute starts before recording but ends during it
-        var samples = [Float](repeating: 1.0, count: 100)
-        let timeline = [
-            MuteTransition(timestamp: 5.0, isMuted: true),
-            MuteTransition(timestamp: 5.8, isMuted: false),
-        ]
-
-        AudioMixer.applyMuteMask(
-            samples: &samples,
-            timeline: timeline,
-            sampleRate: 100,
-            micDelay: 0.5,
-            recordingStart: 5.0,
-        )
-
-        // Mute range: (5.0 - 5.0 - 0.5) to (5.8 - 5.0 - 0.5) = -0.5s to 0.3s
-        // Clamped start to 0; end ≈ sample 29-30 (floating point)
-        // Verify most of the range is muted and tail is untouched
-        for i in 0 ..< 29 {
-            XCTAssertEqual(samples[i], 0.0, "Sample \(i) should be muted")
-        }
-        for i in 31 ..< 100 {
-            XCTAssertEqual(samples[i], 1.0, "Sample \(i) should be untouched")
-        }
-    }
-
-    func testMuteMaskEmptyTimeline() {
-        var samples: [Float] = [1.0, 2.0, 3.0]
-        AudioMixer.applyMuteMask(
-            samples: &samples,
-            timeline: [],
-            sampleRate: 48000,
-        )
-        XCTAssertEqual(samples, [1.0, 2.0, 3.0])
-    }
-
     // MARK: - Echo Suppression
 
     func testEchoSuppressionSilencesOverlap() {
