--- app/MeetingTranscriber/Tests/MuteDetectorTests.swift
@@ -1,155 +1,0 @@
-// swiftlint:disable single_test_class
-@testable import MeetingTranscriber
-import XCTest
-
-// MARK: - MuteDetector Tests
-
-@MainActor
-final class MuteDetectorTests: XCTestCase {
-    func testInitialState() {
-        let detector = MuteDetector(teamsPID: 1234)
-        XCTAssertFalse(detector.isActive)
-        XCTAssertTrue(detector.timeline.isEmpty)
-    }
-
-    func testMutedPrefixes() {
-        // "Unmute" button → user is muted
-        XCTAssertTrue(MuteDetector.mutedPrefixes.contains("unmute"))
-        XCTAssertTrue(MuteDetector.mutedPrefixes.contains("stummschaltung aufheben"))
-    }
-
-    func testUnmutedPrefixes() {
-        // "Mute" button → user is unmuted
-        XCTAssertTrue(MuteDetector.unmutedPrefixes.contains("mute"))
-        XCTAssertTrue(MuteDetector.unmutedPrefixes.contains("stummschalten"))
-    }
-
-    func testTimelineRecordingWithMockProvider() {
-        let detector = MuteDetector(teamsPID: 1234, pollInterval: 0.05)
-
-        var callCount = 0
-        detector.muteStateProvider = { _ in
-            callCount += 1
-            // Alternate: unmuted, muted, unmuted
-            switch callCount {
-            case 1: return false
-            case 2: return true
-            case 3: return false
-            default: return nil
-            }
-        }
-
-        // Manually simulate polling without starting the async task
-        // (AXIsProcessTrusted would fail in test environment)
-        // Instead, test the readMuteState logic directly via prefixes
-        XCTAssertEqual(MuteDetector.mutedPrefixes.count, 2)
-        XCTAssertEqual(MuteDetector.unmutedPrefixes.count, 2)
-    }
-
-    func testStopWithoutStart() {
-        let detector = MuteDetector(teamsPID: 1234)
-        // Should not crash
-        detector.stop()
-        XCTAssertFalse(detector.isActive)
-    }
-}
-
-// MARK: - ParticipantReader Filter Tests
-
-final class ParticipantFilterTests: XCTestCase {
-    func testFiltersEmptyStrings() {
-        let result = ParticipantReader.filterParticipantNames(["", "  ", ""])
-        XCTAssertTrue(result.isEmpty)
-    }
-
-    func testFiltersSingleCharacters() {
-        let result = ParticipantReader.filterParticipantNames(["A", "B", "John"])
-        XCTAssertEqual(result, ["John"])
-    }
-
-    func testFiltersNumbers() {
-        let result = ParticipantReader.filterParticipantNames(["123", "456", "Alice"])
-        XCTAssertEqual(result, ["Alice"])
-    }
-
-    func testFiltersTimestamps() {
-        let result = ParticipantReader.filterParticipantNames(["10:30", "2:15", "Bob"])
-        XCTAssertEqual(result, ["Bob"])
-    }
-
-    func testFiltersUILabels() {
-        let result = ParticipantReader.filterParticipantNames([
-            "Mute", "Camera", "Share", "Alice", "Bob",
-        ])
-        XCTAssertEqual(result, ["Alice", "Bob"])
-    }
-
-    func testFiltersUILabelsCase() {
-        let result = ParticipantReader.filterParticipantNames([
-            "mute", "CAMERA", "Share", "Alice",
-        ])
-        XCTAssertEqual(result, ["Alice"])
-    }
-
-    func testRemovesYouSuffix() {
-        let result = ParticipantReader.filterParticipantNames(["Alice (you)", "Bob"])
-        XCTAssertEqual(result, ["Alice", "Bob"])
-    }
-
-    func testDeduplicates() {
-        let result = ParticipantReader.filterParticipantNames(["Alice", "Bob", "Alice"])
-        XCTAssertEqual(result, ["Alice", "Bob"])
-    }
-
-    func testFiltersParticipantsLabel() {
-        let result = ParticipantReader.filterParticipantNames([
-            "Participants", "In this meeting", "Alice", "Bob",
-        ])
-        XCTAssertEqual(result, ["Alice", "Bob"])
-    }
-
-    func testFiltersRoleLabels() {
-        let result = ParticipantReader.filterParticipantNames([
-            "Presenter", "Attendee", "Organizer", "Alice",
-        ])
-        XCTAssertEqual(result, ["Alice"])
-    }
-
-    func testPreservesNormalNames() {
-        let names = ["John Doe", "Jane Smith", "Max Mustermann"]
-        let result = ParticipantReader.filterParticipantNames(names)
-        XCTAssertEqual(result, names)
-    }
-
-    func testFiltersGermanLabels() {
-        // "stummschalten" starts with a skip pattern
-        let result = ParticipantReader.filterParticipantNames([
-            "Stummschalten", "Alice",
-        ])
-        // "stummschalten" is in skipPatterns (lowercase match)
-        // Actually it's not directly - but it starts with "mute"? No.
-        // Let me check: skipPatterns contains "mute" and stummschalten doesn't start with "mute"
-        // So it would NOT be filtered by skipPatterns
-        // But we have unmutedPrefixes in MuteDetector, not in skipPatterns
-        // filterParticipantNames only checks skipPatterns, not mute prefixes
-        XCTAssertEqual(result, ["Stummschalten", "Alice"])
-    }
-
-    func testRealWorldTeamsRoster() {
-        // Simulated Teams roster with UI clutter
-        let texts = [
-            "People",
-            "In this meeting",
-            "4",
-            "Alice Johnson (you)",
-            "Bob Smith",
-            "Charlie Brown",
-            "Diana Prince",
-            "Mute",
-            "Camera",
-            "Raise hand",
-        ]
-        let result = ParticipantReader.filterParticipantNames(texts)
-        XCTAssertEqual(result, ["Alice Johnson", "Bob Smith", "Charlie Brown", "Diana Prince"])
-    }
-}
